+++ crates/client/gateway/client/src/methods.rs
@@ -1,19 +1,28 @@
-use super::{builder::FeederClient, request_builder::RequestBuilder};
-use crate::error::{SequencerError, StarknetError};
+use std::{borrow::Cow, sync::Arc};
+
 use mp_block::{BlockId, BlockTag};
 use mp_class::{ContractClass, FlattenedSierraClass};
+use mp_gateway::error::{SequencerError, StarknetError};
 use mp_gateway::{
     block::{ProviderBlock, ProviderBlockPending, ProviderBlockPendingMaybe, ProviderBlockSignature},
     state_update::{
         ProviderStateUpdate, ProviderStateUpdatePending, ProviderStateUpdatePendingMaybe, ProviderStateUpdateWithBlock,
         ProviderStateUpdateWithBlockPending, ProviderStateUpdateWithBlockPendingMaybe,
     },
+    user_transaction::{
+        UserDeclareTransaction, UserDeployAccountTransaction, UserInvokeFunctionTransaction, UserTransaction,
+    },
 };
+use serde::de::DeserializeOwned;
 use serde_json::Value;
-use starknet_core::types::{contract::legacy::LegacyContractClass, Felt};
-use std::{borrow::Cow, sync::Arc};
+use starknet_core::types::{
+    contract::legacy::LegacyContractClass, DeclareTransactionResult, DeployAccountTransactionResult, Felt,
+    InvokeTransactionResult,
+};
 
-impl FeederClient {
+use super::{builder::GatewayProvider, request_builder::RequestBuilder};
+
+impl GatewayProvider {
     pub async fn get_block(&self, block_id: BlockId) -> Result<ProviderBlockPendingMaybe, SequencerError> {
         let request = RequestBuilder::new(&self.client, self.feeder_gateway_url.clone(), self.headers.clone())
             .add_uri_segment("get_block")
@@ -99,6 +108,38 @@ impl FeederClient {
             Err(SequencerError::DeserializeBody { serde_error: err })
         }
     }
+
+    async fn add_transaction<T>(&self, transaction: UserTransaction) -> Result<T, SequencerError>
+    where
+        T: DeserializeOwned,
+    {
+        let request = RequestBuilder::new(&self.client, self.gateway_url.clone(), self.headers.clone())
+            .add_uri_segment("add_transaction")
+            .expect("Failed to add URI segment. This should not fail in prod.");
+
+        request.send_post(transaction).await
+    }
+
+    pub async fn add_invoke_transaction(
+        &self,
+        transaction: UserInvokeFunctionTransaction,
+    ) -> Result<InvokeTransactionResult, SequencerError> {
+        self.add_transaction(UserTransaction::InvokeFunction(transaction)).await
+    }
+
+    pub async fn add_declare_transaction(
+        &self,
+        transaction: UserDeclareTransaction,
+    ) -> Result<DeclareTransactionResult, SequencerError> {
+        self.add_transaction(UserTransaction::Declare(transaction)).await
+    }
+
+    pub async fn add_deploy_account_transaction(
+        &self,
+        transaction: UserDeployAccountTransaction,
+    ) -> Result<DeployAccountTransactionResult, SequencerError> {
+        self.add_transaction(UserTransaction::DeployAccount(transaction)).await
+    }
 }
 
 #[cfg(test)]
@@ -110,6 +151,7 @@ mod tests {
     };
     use mp_block::BlockTag;
     use mp_class::CompressedLegacyContractClass;
+    use mp_gateway::error::{SequencerError, StarknetError, StarknetErrorCode};
     use rstest::*;
     use serde::de::DeserializeOwned;
     use starknet_core::types::Felt;
@@ -117,8 +159,6 @@ mod tests {
     use std::io::{BufReader, BufWriter, Read, Write};
     use std::ops::Drop;
     use std::path::PathBuf;
-
-    use crate::error::StarknetErrorCode;
 
     use super::*;
 
@@ -211,7 +251,7 @@ mod tests {
     #[test]
     #[ignore]
     fn compress() {
-        // file_compress("src/client/mocks/.json").unwrap();
+        // file_compress("src/mocks/.json").unwrap();
     }
 
     /// Converts a crate-relative path to an absolute system path.
@@ -225,13 +265,13 @@ mod tests {
     }
 
     #[fixture]
-    fn client_mainnet_fixture() -> FeederClient {
-        FeederClient::starknet_alpha_mainnet()
+    fn client_mainnet_fixture() -> GatewayProvider {
+        GatewayProvider::starknet_alpha_mainnet()
     }
 
     #[rstest]
     #[tokio::test]
-    async fn get_block(client_mainnet_fixture: FeederClient) {
+    async fn get_block(client_mainnet_fixture: GatewayProvider) {
         let block = client_mainnet_fixture.get_block(BlockId::Number(0)).await.unwrap();
         println!("parent_block_hash: 0x{:x}", block.parent_block_hash());
         assert!(matches!(block, ProviderBlockPendingMaybe::NonPending(_)));
@@ -253,7 +293,7 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_state_update(client_mainnet_fixture: FeederClient) {
+    async fn get_state_update(client_mainnet_fixture: GatewayProvider) {
         let state_update = client_mainnet_fixture.get_state_update(BlockId::Number(0)).await.unwrap();
         assert!(matches!(state_update, ProviderStateUpdatePendingMaybe::NonPending(_)));
         assert_eq!(
@@ -284,7 +324,7 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_state_update_with_block_first_few_blocks(client_mainnet_fixture: FeederClient) {
+    async fn get_state_update_with_block_first_few_blocks(client_mainnet_fixture: GatewayProvider) {
         let let_binding = client_mainnet_fixture
             .get_state_update_with_block(BlockId::Number(0))
             .await
@@ -294,7 +334,7 @@ mod tests {
             state_update_0.non_pending_ownded().expect("State update at block number 0 should not be pending");
         let block_0 = block_0.non_pending_owned().expect("Block at block number 0 should not be pending");
         let ProviderStateUpdateWithBlock { state_update: state_update_0_reference, block: block_0_reference } =
-            load_from_file_compressed::<ProviderStateUpdateWithBlock>("src/client/mocks/state_update_and_block_0.gz");
+            load_from_file_compressed::<ProviderStateUpdateWithBlock>("src/mocks/state_update_and_block_0.gz");
 
         assert_eq!(state_update_0, state_update_0_reference);
         assert_eq!(block_0, block_0_reference);
@@ -308,7 +348,7 @@ mod tests {
             state_update_1.non_pending_ownded().expect("State update at block number 1 should not be pending");
         let block_1 = block_1.non_pending_owned().expect("Block at block number 1 should not be pending");
         let ProviderStateUpdateWithBlock { state_update: state_update_1_reference, block: block_1_reference } =
-            load_from_file_compressed::<ProviderStateUpdateWithBlock>("src/client/mocks/state_update_and_block_1.gz");
+            load_from_file_compressed::<ProviderStateUpdateWithBlock>("src/mocks/state_update_and_block_1.gz");
 
         assert_eq!(state_update_1, state_update_1_reference);
         assert_eq!(block_1, block_1_reference);
@@ -322,7 +362,7 @@ mod tests {
             state_update_2.non_pending_ownded().expect("State update at block number 0 should not be pending");
         let block_2 = block_2.non_pending_owned().expect("Block at block number 0 should not be pending");
         let ProviderStateUpdateWithBlock { state_update: state_update_2_reference, block: block_2_reference } =
-            load_from_file_compressed::<ProviderStateUpdateWithBlock>("src/client/mocks/state_update_and_block_2.gz");
+            load_from_file_compressed::<ProviderStateUpdateWithBlock>("src/mocks/state_update_and_block_2.gz");
 
         assert_eq!(state_update_2, state_update_2_reference);
         assert_eq!(block_2, block_2_reference);
@@ -330,7 +370,7 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_state_update_with_block_latest(client_mainnet_fixture: FeederClient) {
+    async fn get_state_update_with_block_latest(client_mainnet_fixture: GatewayProvider) {
         let let_binding = client_mainnet_fixture
             .get_state_update_with_block(BlockId::Tag(BlockTag::Latest))
             .await
@@ -343,7 +383,7 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_state_update_with_block_pending(client_mainnet_fixture: FeederClient) {
+    async fn get_state_update_with_block_pending(client_mainnet_fixture: GatewayProvider) {
         let let_binding = client_mainnet_fixture
             .get_state_update_with_block(BlockId::Tag(BlockTag::Pending))
             .await
@@ -356,14 +396,13 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_class_by_hash_block_0(client_mainnet_fixture: FeederClient) {
+    async fn get_class_by_hash_block_0(client_mainnet_fixture: GatewayProvider) {
         let class = client_mainnet_fixture
             .get_class_by_hash(Felt::from_hex_unchecked(CLASS_BLOCK_0), BlockId::Number(0))
             .await
             .unwrap_or_else(|_| panic!("Getting class {CLASS_BLOCK_0} at block number 0"));
-        let class_reference = load_from_file_compressed::<LegacyContractClass>(&format!(
-            "src/client/mocks/class_block_0_{CLASS_BLOCK_0}.gz"
-        ));
+        let class_reference =
+            load_from_file_compressed::<LegacyContractClass>(&format!("src/mocks/class_block_0_{CLASS_BLOCK_0}.gz"));
         let class_compressed_reference: CompressedLegacyContractClass =
             class_reference.compress().expect("Compressing legacy contract class").into();
 
@@ -372,13 +411,13 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_class_by_hash_account(client_mainnet_fixture: FeederClient) {
+    async fn get_class_by_hash_account(client_mainnet_fixture: GatewayProvider) {
         let class_account = client_mainnet_fixture
             .get_class_by_hash(Felt::from_hex_unchecked(CLASS_ACCOUNT), BlockId::Number(CLASS_ACCOUNT_BLOCK))
             .await
             .unwrap_or_else(|_| panic!("Getting account class {CLASS_ACCOUNT} at block number {CLASS_ACCOUNT_BLOCK}"));
         let class_reference = load_from_file_compressed::<LegacyContractClass>(&format!(
-            "src/client/mocks/class_block_{CLASS_ACCOUNT_BLOCK}_account_{CLASS_ACCOUNT}.gz"
+            "src/mocks/class_block_{CLASS_ACCOUNT_BLOCK}_account_{CLASS_ACCOUNT}.gz"
         ));
         let class_compressed_reference: CompressedLegacyContractClass =
             class_reference.compress().expect("Compressing legacy contract class").into();
@@ -388,13 +427,13 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_class_by_hash_proxy(client_mainnet_fixture: FeederClient) {
+    async fn get_class_by_hash_proxy(client_mainnet_fixture: GatewayProvider) {
         let class_proxy = client_mainnet_fixture
             .get_class_by_hash(Felt::from_hex_unchecked(CLASS_PROXY), BlockId::Number(CLASS_PROXY_BLOCK))
             .await
             .unwrap_or_else(|_| panic!("Getting proxy class {CLASS_PROXY} at block number {CLASS_PROXY_BLOCK}"));
         let class_reference = load_from_file_compressed::<LegacyContractClass>(&format!(
-            "src/client/mocks/class_block_{CLASS_PROXY_BLOCK}_proxy_{CLASS_PROXY}.gz"
+            "src/mocks/class_block_{CLASS_PROXY_BLOCK}_proxy_{CLASS_PROXY}.gz"
         ));
         let class_compressed_reference: CompressedLegacyContractClass =
             class_reference.compress().expect("Compressing legacy contract class").into();
@@ -404,13 +443,13 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_class_by_hash_erc20(client_mainnet_fixture: FeederClient) {
+    async fn get_class_by_hash_erc20(client_mainnet_fixture: GatewayProvider) {
         let class_erc20 = client_mainnet_fixture
             .get_class_by_hash(Felt::from_hex_unchecked(CLASS_ERC20), BlockId::Number(CLASS_ERC20_BLOCK))
             .await
             .unwrap_or_else(|_| panic!("Getting proxy class {CLASS_ERC20} at block number {CLASS_ERC20_BLOCK}"));
         let class_reference = load_from_file_compressed::<LegacyContractClass>(&format!(
-            "src/client/mocks/class_block_{CLASS_ERC20_BLOCK}_erc20_{CLASS_ERC20}.gz"
+            "src/mocks/class_block_{CLASS_ERC20_BLOCK}_erc20_{CLASS_ERC20}.gz"
         ));
         let class_compressed_reference: CompressedLegacyContractClass =
             class_reference.compress().expect("Compressing legacy contract class").into();
@@ -420,13 +459,13 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_class_by_hash_erc721(client_mainnet_fixture: FeederClient) {
+    async fn get_class_by_hash_erc721(client_mainnet_fixture: GatewayProvider) {
         let class_erc721 = client_mainnet_fixture
             .get_class_by_hash(Felt::from_hex_unchecked(CLASS_ERC721), BlockId::Number(CLASS_ERC721_BLOCK))
             .await
             .unwrap_or_else(|_| panic!("Getting proxy class {CLASS_ERC721} at block number {CLASS_ERC721_BLOCK}"));
         let class_reference = load_from_file_compressed::<LegacyContractClass>(&format!(
-            "src/client/mocks/class_block_{CLASS_ERC721_BLOCK}_erc721_{CLASS_ERC721}.gz"
+            "src/mocks/class_block_{CLASS_ERC721_BLOCK}_erc721_{CLASS_ERC721}.gz"
         ));
         let class_compressed_reference: CompressedLegacyContractClass =
             class_reference.compress().expect("Compressing legacy contract class").into();
@@ -436,13 +475,13 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_class_by_hash_erc1155(client_mainnet_fixture: FeederClient) {
+    async fn get_class_by_hash_erc1155(client_mainnet_fixture: GatewayProvider) {
         let class_erc1155 = client_mainnet_fixture
             .get_class_by_hash(Felt::from_hex_unchecked(CLASS_ERC1155), BlockId::Number(CLASS_ERC1155_BLOCK))
             .await
             .unwrap_or_else(|_| panic!("Getting proxy class {CLASS_ERC1155} at block number {CLASS_ERC1155_BLOCK}"));
         let class_reference = load_from_file_compressed::<LegacyContractClass>(&format!(
-            "src/client/mocks/class_block_{CLASS_ERC1155_BLOCK}_erc1155_{CLASS_ERC1155}.gz"
+            "src/mocks/class_block_{CLASS_ERC1155_BLOCK}_erc1155_{CLASS_ERC1155}.gz"
         ));
         let class_compressed_reference: CompressedLegacyContractClass =
             class_reference.compress().expect("Compressing legacy contract class").into();
@@ -452,13 +491,12 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_signature_first_few_blocks(client_mainnet_fixture: FeederClient) {
+    async fn get_signature_first_few_blocks(client_mainnet_fixture: GatewayProvider) {
         let signature_block_0 = client_mainnet_fixture
             .get_signature(BlockId::Number(0))
             .await
             .expect("Failed to get signature for block number 0");
-        let signature_reference =
-            load_from_file_compressed::<ProviderBlockSignature>("src/client/mocks/signature_block_0.gz");
+        let signature_reference = load_from_file_compressed::<ProviderBlockSignature>("src/mocks/signature_block_0.gz");
 
         assert_eq!(signature_block_0, signature_reference);
 
@@ -466,8 +504,7 @@ mod tests {
             .get_signature(BlockId::Number(1))
             .await
             .expect("Failed to get signature for block number 1");
-        let signature_reference =
-            load_from_file_compressed::<ProviderBlockSignature>("src/client/mocks/signature_block_1.gz");
+        let signature_reference = load_from_file_compressed::<ProviderBlockSignature>("src/mocks/signature_block_1.gz");
 
         assert_eq!(signature_block_1, signature_reference);
 
@@ -475,15 +512,14 @@ mod tests {
             .get_signature(BlockId::Number(2))
             .await
             .expect("Failed to get signature for block number 2");
-        let signature_reference =
-            load_from_file_compressed::<ProviderBlockSignature>("src/client/mocks/signature_block_2.gz");
+        let signature_reference = load_from_file_compressed::<ProviderBlockSignature>("src/mocks/signature_block_2.gz");
 
         assert_eq!(signature_block_2, signature_reference);
     }
 
     #[rstest]
     #[tokio::test]
-    async fn get_signature_latest(client_mainnet_fixture: FeederClient) {
+    async fn get_signature_latest(client_mainnet_fixture: GatewayProvider) {
         let signature_block_latest = client_mainnet_fixture.get_signature(BlockId::Tag(BlockTag::Latest)).await;
 
         assert!(matches!(signature_block_latest, Ok(ProviderBlockSignature { .. })))
@@ -491,7 +527,7 @@ mod tests {
 
     #[rstest]
     #[tokio::test]
-    async fn get_signature_pending(client_mainnet_fixture: FeederClient) {
+    async fn get_signature_pending(client_mainnet_fixture: GatewayProvider) {
         let signature_block_pending = client_mainnet_fixture.get_signature(BlockId::Tag(BlockTag::Pending)).await;
 
         assert!(matches!(