+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - feat(endpoint): added extra admin rpc endpoint for sensitive rpc calls
+- feat(fgw): added `add_transaction` for gateway client
 - fix(fgw): include `l1_to_l2_consumed_message` in L1 handler receipt
 - build: up starknet-rs, starknet-types, blockifier(v0.8.0), cairo
 - feat(rpc): added `getCompiledCasm` method